refining spy tests
