+++ src/kernel/CMakeLists.txt
@@ -21,7 +21,6 @@ race.test.c
 spellbook.test.c
 curse.test.c
 jsonconf.test.c
-spy.test.c
 )
 
 SET(_FILES
